Move `CurveGeometryMismatch` to `validation`
